refactor: 너무 큰 용량의 svg들 모두 png로 변경(js 번들 크기에 영향)
