+++ src/assets/images/locationSharingPage/arrivalPin.png
