+++ src/pages/NotificationPage/components/NotificationItem.tsx
@@ -1,11 +1,11 @@
 import { useNavigate } from "react-router-dom";
 import styles from "./notificationItem.module.scss";
-import DefaultNotificationIcon from "@assets/Icons/alert/defaultAlert.svg?react";
 import RedDotIcon from "@assets/Icons/headers/redDot.svg?react";
 import { usePostReadNotification } from "@api/notification/postReadNotification";
 import useAuthStore from "@store/useAuthStore";
 import { notificationType } from "../utils/notificationType";
 import { getRelativeTime } from "../utils/getRelativeTime";
+import defaultAlertIcon from "@assets/images/notificationPage/defaultAlert.png"
 
 interface Props {
   notificationItem: INotificationItemType;
@@ -26,7 +26,7 @@ const NotificationItem: React.FC<Props> = ({ notificationItem }) => {
   return (
     <div className={styles.notificationItemContainer} onClick={handleNotificationClick}>
       <div className={styles.leftSection}>
-        <DefaultNotificationIcon width={53} height={51} />
+        <img src={defaultAlertIcon} width={53} height={51} />
       </div>
       <div className={styles.rightSection}>
         <div className={styles.eventTypeText}>{notificationType(notificationItem.eventType)}</div>
