+++ src/pages/RegisterSuccessPage/page/index.tsx
@@ -1,19 +1,19 @@
-import RegisterSuccessIcon from "@assets/Icons/registerSuccessPage/registerSuccessIcon.svg?react"
-import styles from "./registerSuccess.module.scss"
-import OnlyTextHeader from "@components/headers/OnlyTextHeader"
-import { useNavigate } from "react-router-dom"
-import LoginButton from "@components/buttons/LoginButton"
+import registerSuccessImage from "@assets/images/registerSuccessPage/registerSuccess.png";
+import styles from "./registerSuccess.module.scss";
+import OnlyTextHeader from "@components/headers/OnlyTextHeader";
+import { useNavigate } from "react-router-dom";
+import LoginButton from "@components/buttons/LoginButton";
 
 const RegisterSuccessPage = () => {
   const navigate = useNavigate();
   const handleGoLogin = () => {
-    navigate('/login')
-  }
+    navigate("/login");
+  };
   return (
     <div className={styles.mainContainer}>
       <OnlyTextHeader title="가입 완료" backgroundColor="purple" />
       <div className={styles.centerSection}>
-        <RegisterSuccessIcon width={210} height={206} />
+        <img src={registerSuccessImage} width={210} height={206} />
         <div className={styles.bigText}>가입이 완료되었습니다</div>
         <div className={styles.smallText}>
           <div>PlanU에서 모든 일정을 확인하고</div>
@@ -25,6 +25,6 @@ const RegisterSuccessPage = () => {
       </div>
     </div>
   );
-}
+};
 
-export default RegisterSuccessPage
+export default RegisterSuccessPage;
